refactor: removed extra interfaces, test corrected so, that _isMounted extra variable doesn't be needed any more
