+++ react.components/src/components/SearchBar/SearchBar.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import SearchBar from '../components/SearchBar/SearchBar';
+import SearchBar from './SearchBar';
 import userEvent from '@testing-library/user-event/dist';
-import fakeLocalStorage from './functions/fackeLocalStorage';
+import fakeLocalStorage from '../../tests/functions/fackeLocalStorage';
 
 describe('SearchBar testing', () => {
   beforeEach(() => {
